% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNH.R
\name{CNH}
\alias{CNH}
\title{copy number intra-tumor heterogeneity (CNH) from a single copy number measurement}
\usage{
CNH(seg_val,seg_len,ploidy,purity)
}
\arguments{
\item{seg_val}{numeric vector with values of relative copy numbers per segment.}

\item{seg_len}{numeric vector with segment lengths.}

\item{ploidy}{tumour ploidy}

\item{purity}{sample purity, default = NULL for grid search}
}
\value{
1st output argument: \code{CNH_out}      --> inferred CNH

2nd output argument: \code{ploidy_out}  -->  inferred ploidy for empty input ploidy, otherwise same as input ploidy.

3th output argument: \code{purity_out}   -->  inferred purity for empty input purity, otherwise same as input purity.
}
\description{
MATLAB CNH.m function translation into R to calculate copy number intra-tumor heterogeneity (CNH) from a single copy number measurement.
In this method, relative segmented copy numbers are transformed to absolute copy numbers and the distance
to integer values is measured for each segment. This procedure is
performed for either a range of ploidies/purities or for fixed
ploidy/purity, in case either or both quantities are known from previous
measurements.
Copy number heterogeneity (CNH) is defined as the minimum average distance of
segments to the closest integer value, weighted by segment length.
In addition to CNH, this function returns the ploidy and purity
corresponding to the inferred CNH.
}
\examples{
file = paste0(path.package("CNH"), "/SJBALL247_D.bed")
SJBALL247_D <- read.delim(file, header=FALSE)
seg_len <- SJBALL247_D[,3]-SJBALL247_D[,2]
seg_val<- SJBALL247_D[,4]
CNH(seg_val,seg_len)
}
